refactor: type 패키지 이름을 constant로 변경
