+++ src/main/java/baseball/constant/ResultTurn.java
@@ -1,4 +1,4 @@
-package baseball.type;
+package baseball.constant;
 
 public enum ResultTurn {
     Nothing,
